requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.34.113
botocore==1.34.113
certifi==2024.2.2
charset-normalizer==3.3.2
cohere==5.5.3
docx==0.2.4
exceptiongroup==1.2.1
faiss-cpu==1.8.0
fastavro==1.9.4
filelock==3.14.0
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.1
idna==3.7
jmespath==1.0.1
lxml==5.2.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.12.0
tzdata==2024.1
urllib3==1.26.18
